backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
Pillow==11.1.0
